Fixed #418 - Windows State/Size does not save in case of fullscreen
It seems to be a QT issue. The workaround described in the issue
QTBUG-46620 was applied, but not everywhere restoreGeometry is called, as it caused
other issues in QT, which are also detailed in the bug report. While it works better,
it is not perfect.

Bug QTBUG-46620 was fixed in QT 6.3.0, so it is necessary to determine how this workaround
behaves under this QT version. This could affect Windows users.