Fixing #455 - Force set encoding for ADI formatter to Latin1
In cases when an ADI record contains non-ASCII characters,
Qt internally tries to interpret them as UTF-8 and convert them,
which can lead to unpredictable behavior. This commit fixes the issue #455
as a side effect. Unfortunately, if the ADIF contains non-ASCII characters,
QLog is still unable to handle them correctly - what is not an issue because
ADI-format is defined as ASCII format.