Fixing #472 - HamlibDrv: Added a postponed handling for Rig soft errors
In some cases, the rig may be busy, which can lead to an error message
and a disconnect event. However, this does not mean that communication has stopped.
This commit implements postponed error handling with a delay of 15 seconds.
If the same error does not occur during this time and there is a successful
operation for this type of error, then the error message is not sent
to the operator. Currently, error handling is performed this way only
for 'Get Freq' and 'Get Mode'.

QLog also recognizes two types of errors: hard and soft.
The hard error is reported immediately. The soft error is postponed.