POTA Export: Code tuning; Export file; Division of export files
The export generates two types of files. The first type contains
all QSOs required for reporting as a Hunter. This file has the same
name as the one entered in the export window. The second type of file
contains QSOs as an Activator. These QSOs are split into files based
on the POTA Ref ID. In both cases, the QSOs are expanded by POTA Ref
and the SIG and SIG_INFO fields are populated.