ADIF315: Stopped exporting fields with the default value N
There are fields with a default value of 'N'.
Therefore, it is not necessary to explicitly export them
when the local database also has the value 'N'.
This approach can be helpful, for example, in the case of new fields
in ADIF, where external systems are not yet prepared, and QLog would
generate ADIF with the default value 'N'.