Fixed #608 - Logbook Country translation
This fix adds the ability to translate the Contry_INTL fields.
One is only about translation, because the fields are free-text fields.
Therefore, we cannot insert a combo with countries here, because some country
may be imported that is not in this combo. This fix leaves the ability for the operator
to write any string. If the country is in English, it will be translated.