Fixed #633 - QSO Detail show QSLMSG_RCVD field
Issue1:
Originally, QSLMSG was a general QSL field where QSL messages were exchanged.
Now, this field represents sending. But older entries may contain messages that were received.
The problem is that it is not possible migrate the QSLMSG field to QSLMSG_RCVD because
we don't know whether it contains sent or received messages.

Newly, QSO Detail displayes QSLMSG_RCVD therefore old received QSL messages, which are stored in QSLMSG
may not be displayed.

Issue2:
There is no option how to edit QSLMSG in QLog - message which will be send out. Since we don't know
whether QSLMSG contain old received message a new sent message, it's hard to decide whether
to allow this field to be edited.

There is still an option to edit QSLMSG directly in Logbook, so we are leaving it unimplemented for now.